+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`0.6.0...master`](https://github.com/localheinz/phpstan-rules/compare/0.6.0...master).
 
+### Added
+
+* Added `Classes\NoExtendsRule`, which reports an error when a class extends a class that is not allowed to be extended ([#68](https://github.com/localheinz/phpstan-rules/pull/68)), by [@localheinz](https://github.com/localheinz)
+
 ## [`0.6.0`](https://github.com/localheinz/phpstan-rules/releases/tag/0.6.0)
 
 For a full diff see [`0.5.0...0.6.0`](https://github.com/localheinz/phpstan-rules/compare/0.5.0...0.6.0).